Update cr_reject.py
I was hitting this error in the script:

"ValueError: Input file is ifj410xhq_flt, but only FLT or FLC files are valid."

this is because the string was being split on a hyphen "-" instead of an underscore "_". 

This PR fixes that, and it correctly gets the file type as flt. :)